+++ tasks/msteams/src/main/java/com/walmartlabs/concord/plugins/msteams/Constants.java
@@ -41,15 +41,27 @@ public class Constants {
     public static final String MESSAGE_SECTIONS_KEY = "sections";
     public static final String MESSAGE_POTENTIAL_ACTION_KEY = "potentialAction";
     public static final String IGNORE_ERRORS_KEY = "ignoreErrors";
+    public static final String VAR_ACCESS_TOKEN_API = "accessTokenApi";
+    public static final String VAR_ACCESS_TOKEN = "access_token";
 
     public static final String PROXY_ADDRESS_KEY = "proxyAddress";
     public static final String PROXY_PORT_KEY = "proxyPort";
     public static final String CONNECTION_TIMEOUT_KEY = "connectTimeout";
     public static final String SO_TIMEOUT_KEY = "soTimeout";
     public static final String RETRY_COUNT_KEY = "retryCount";
+    public static final String API_BOT_FRAMEWORK_SCOPE = "https://api.botframework.com/.default";
+    public static final String API_BOT_FRAMEWORK_GRANT_TYPE = "client_credentials";
+    public static final String VAR_CLIENT_ID = "clientId";
+    public static final String VAR_CLIENT_SECRET = "clientSecret";
+    public static final String VAR_CHANNEL_ID = "channelId";
+    public static final String VAR_ACTIVITY = "activity";
+    public static final String VAR_ROOT_API = "rootApi";
+    public static final String USE_PROXY_KEY = "useProxy";
+    public static final String VAR_CONVERSATION_ID = "conversationId";
 
     public static final int TOO_MANY_REQUESTS_ERROR = 429;
     public static final int TEAMS_SUCCESS_STATUS_CODE = 200;
+    public static final int TEAMS_SUCCESS_STATUS_CODE_V2 = 201;
     public static final int DEFAULT_RETRY_AFTER = 1;
     public static final int DEFAULT_PROXY_PORT = 8080;
 
@@ -63,7 +75,13 @@ public class Constants {
             Constants.MESSAGE_THEME_COLOR_KEY,
             Constants.MESSAGE_SECTIONS_KEY,
             Constants.MESSAGE_POTENTIAL_ACTION_KEY,
-            Constants.IGNORE_ERRORS_KEY
+            Constants.IGNORE_ERRORS_KEY,
+            Constants.VAR_CHANNEL_ID,
+            Constants.VAR_ACTIVITY,
+            Constants.VAR_CLIENT_ID,
+            Constants.VAR_CLIENT_SECRET,
+            Constants.USE_PROXY_KEY,
+            Constants.VAR_CONVERSATION_ID
     };
 
     private Constants() {